+++ app/sprinkles/core/src/Core.php
@@ -33,9 +33,7 @@ class Core extends Sprinkle
     {
         $this->ci = $ci;
 
-        // Register core locator streams
-        $this->ci->locator->registerStream('migrations', '', \UserFrosting\MIGRATIONS_DIR);
-        $this->ci->locator->registerStream('seeds', '', \UserFrosting\SEEDS_DIR);
+        $this->registerStreams();
     }
 
     /**
@@ -113,5 +111,30 @@ class Core extends Sprinkle
     public function onAddGlobalMiddleware(Event $event)
     {
         SlimCsrfProvider::registerMiddleware($event->getApp(), $this->ci->request, $this->ci->csrf);
+    }
+
+    /**
+     * Register Core sprinkle locator streams
+     */
+    protected function registerStreams()
+    {
+        /** @var \UserFrosting\UniformResourceLocator\ResourceLocator $locator */
+        $locator = $this->ci->locator;
+
+        // Register core locator shared streams
+        $locator->registerStream('cache', '', \UserFrosting\APP_DIR_NAME . \UserFrosting\DS . \UserFrosting\CACHE_DIR_NAME, true);
+        $locator->registerStream('config', '', \UserFrosting\DS . \UserFrosting\CONFIG_DIR_NAME);
+        $locator->registerStream('log', '', \UserFrosting\APP_DIR_NAME . \UserFrosting\DS . \UserFrosting\LOG_DIR_NAME, true);
+        $locator->registerStream('migrations', '', \UserFrosting\MIGRATIONS_DIR);
+        $locator->registerStream('seeds', '', \UserFrosting\SEEDS_DIR);
+        $locator->registerStream('session', '', \UserFrosting\APP_DIR_NAME . \UserFrosting\DS . \UserFrosting\SESSION_DIR_NAME, true);
+
+        // Register core locator sprinkle streams
+        $locator->registerStream('extra', '', \UserFrosting\DS . \UserFrosting\EXTRA_DIR_NAME);
+        $locator->registerStream('factories', '', \UserFrosting\DS . \UserFrosting\FACTORY_DIR_NAME);
+        $locator->registerStream('locale', '', \UserFrosting\DS . \UserFrosting\LOCALE_DIR_NAME);
+        $locator->registerStream('routes', '', \UserFrosting\DS . \UserFrosting\ROUTE_DIR_NAME);
+        $locator->registerStream('schema', '', \UserFrosting\DS . \UserFrosting\SCHEMA_DIR_NAME);
+        $locator->registerStream('templates', '', \UserFrosting\DS . \UserFrosting\TEMPLATE_DIR_NAME);
     }
 }